Prep range3d <> range3d encodings
