+++ data-model/src/relative_encodings/mod.rs
@@ -1,14 +1,11 @@
 mod path_path;
-pub use path_path::*;
 
 mod entry_entry;
-pub use entry_entry::*;
 
 mod entry_area;
-pub use entry_area::*;
 
 mod entry_range3d;
-pub use entry_range3d::*;
 
 mod area_area;
-pub use area_area::*;
+
+mod range3d_range3d;
